pyproject.toml
[project]
name = "numeraz"
version = "0.1.1"
description = "Numerai Model Development with Azure ML"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-ml>=1.22.4",
    "click>=8.1.7",
    "numerapi>=2.19.1",
]

[project.scripts]
naz = "cli:cli"
